Factor conversion, handling of initialization expressions, and pending blocks out of the checker context.
