+++ toolchain/check/handle_if_expression.cpp
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 #include "toolchain/check/context.h"
+#include "toolchain/check/convert.h"
 
 namespace Carbon::Check {
 
@@ -13,7 +14,7 @@ auto HandleIfExpressionIf(Context& context, Parse::Node parse_node) -> bool {
   auto cond_value_id = context.node_stack().PopExpression();
 
   // Convert the condition to `bool`, and branch on it.
-  cond_value_id = context.ConvertToBoolValue(if_node, cond_value_id);
+  cond_value_id = ConvertToBoolValue(context, if_node, cond_value_id);
   auto then_block_id =
       context.AddDominatedBlockAndBranchIf(if_node, cond_value_id);
   auto else_block_id = context.AddDominatedBlockAndBranch(if_node);
@@ -33,7 +34,7 @@ auto HandleIfExpressionThen(Context& context, Parse::Node parse_node) -> bool {
       context.node_stack().Peek<Parse::NodeKind::IfExpressionIf>();
 
   // Convert the first operand to a value.
-  then_value_id = context.ConvertToValueExpression(then_value_id);
+  then_value_id = ConvertToValueExpression(context, then_value_id);
 
   // Start emitting the `else` block.
   context.node_block_stack().Push(else_block_id);
@@ -58,7 +59,7 @@ auto HandleIfExpressionElse(Context& context, Parse::Node parse_node) -> bool {
   // TODO: Find a common type, and convert both operands to it instead.
   auto result_type_id = context.semantics_ir().GetNode(then_value_id).type_id();
   else_value_id =
-      context.ConvertToValueOfType(else_node, else_value_id, result_type_id);
+      ConvertToValueOfType(context, else_node, else_value_id, result_type_id);
 
   // Create a resumption block and branches to it.
   auto chosen_value_id = context.AddConvergenceBlockWithArgAndPush(
