+++ toolchain/check/handle_statement.cpp
@@ -3,13 +3,27 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 #include "toolchain/check/context.h"
+#include "toolchain/check/convert.h"
 #include "toolchain/sem_ir/node.h"
 
 namespace Carbon::Check {
 
+// TODO: Find a better home for this. We'll likely need it for more than just
+// expression statements.
+static auto HandleDiscardedExpression(Context& context, SemIR::NodeId expr_id)
+    -> void {
+  // If we discard an initializing expression, convert it to a value or
+  // reference so that it has something to initialize.
+  auto expr = context.semantics_ir().GetNode(expr_id);
+  Convert(context, expr.parse_node(), expr_id,
+          {.kind = ConversionTarget::Discarded, .type_id = expr.type_id()});
+
+  // TODO: This will eventually need to do some "do not discard" analysis.
+}
+
 auto HandleExpressionStatement(Context& context, Parse::Node /*parse_node*/)
     -> bool {
-  context.HandleDiscardedExpression(context.node_stack().PopExpression());
+  HandleDiscardedExpression(context, context.node_stack().PopExpression());
   return true;
 }
 
@@ -52,10 +66,10 @@ auto HandleReturnStatement(Context& context, Parse::Node parse_node) -> bool {
           .Note(fn_node.parse_node(), ReturnStatementImplicitNote)
           .Emit();
     } else if (callable.return_slot_id.is_valid()) {
-      arg = context.Initialize(parse_node, callable.return_slot_id, arg);
+      arg = Initialize(context, parse_node, callable.return_slot_id, arg);
     } else {
-      arg = context.ConvertToValueOfType(parse_node, arg,
-                                         callable.return_type_id);
+      arg = ConvertToValueOfType(context, parse_node, arg,
+                                 callable.return_type_id);
     }
 
     context.AddNode(SemIR::Node::ReturnExpression::Make(parse_node, arg));
